fix config

+++ wandb/sdk/launch/docker.py
@@ -138,7 +138,12 @@ def build_docker_image(
     )
     copy_code_line = ""
     workdir_line = ""
+    copy_config_line = ""
     workdir = os.path.join("/home/", getpass.getuser())
+    if launch_project.override_config:
+        copy_config_line = "COPY {}/{} {}\n".format(
+            _PROJECT_TAR_ARCHIVE_NAME, _project_spec.DEFAULT_CONFIG_PATH, workdir
+        )
     if copy_code:
         copy_code_line = "COPY {}/ {}\n".format(_PROJECT_TAR_ARCHIVE_NAME, workdir)
         workdir_line = "WORKDIR {}\n".format(workdir)
@@ -150,7 +155,7 @@ def build_docker_image(
 
     dockerfile = (
         "FROM {imagename}\n"
-        "COPY {src_tar}/{config_path} {workdir}"
+        "{copy_config_line}"
         "{copy_code_line}"
         "{workdir_line}"
         "{name_line}"
@@ -164,9 +169,7 @@ def build_docker_image(
         "ENV WANDB_DOCKER={docker_image}"
     ).format(
         imagename=base_image,
-        src_tar=_PROJECT_TAR_ARCHIVE_NAME,
-        config_path=_project_spec.DEFAULT_CONFIG_PATH,
-        workdir=workdir,
+        copy_config_line=copy_config_line,
         copy_code_line=copy_code_line,
         workdir_line=workdir_line,
         name_line=name_line,
@@ -176,8 +179,11 @@ def build_docker_image(
         entity=entity,
         run_id=launch_project.run_id or None,
         docker_image=launch_project.docker_image,
+        config_path=_project_spec.DEFAULT_CONFIG_PATH,
     )
-    build_ctx_path = _create_docker_build_ctx(launch_project.project_dir, dockerfile)
+    build_ctx_path = _create_docker_build_ctx(
+        launch_project.project_dir, dockerfile, launch_project.config_path
+    )
     with open(build_ctx_path, "rb") as docker_build_ctx:
         _logger.info("=== Building docker image %s ===", image_uri)
         #  TODO: replace with shelling out
@@ -266,10 +272,11 @@ def _create_docker_build_ctx(
     try:
         dst_path = os.path.join(directory, "wandb-project-contents")
         shutil.copytree(src=work_dir, dst=dst_path)
-        shutil.copyfile(
-            src=config_path,
-            dst=os.path.join(dst_path, _project_spec.DEFAULT_CONFIG_PATH),
-        )
+        if os.path.exists(config_path):
+            shutil.copyfile(
+                src=config_path,
+                dst=os.path.join(dst_path, _project_spec.DEFAULT_CONFIG_PATH),
+            )
         with open(os.path.join(dst_path, _GENERATED_DOCKERFILE_NAME), "w") as handle:
             handle.write(dockerfile_contents)
         _, result_path = tempfile.mkstemp()
